# api/rest/openapi/types.gen.go
// Package openapi provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen version v1.13.4 DO NOT EDIT.
package openapi

import (
	openapi_types "github.com/deepmap/oapi-codegen/pkg/types"
)

const (
	KeyScopes    = "key.Scopes"
	SecretScopes = "secret.Scopes"
)

// Defines values for LeaseRentInterval.
const (
	LeaseRentIntervalDaily   LeaseRentInterval = "daily"
	LeaseRentIntervalMonthly LeaseRentInterval = "monthly"
	LeaseRentIntervalWeekly  LeaseRentInterval = "weekly"
)

// Defines values for MinLeaseRentInterval.
const (
	MinLeaseRentIntervalDaily   MinLeaseRentInterval = "daily"
	MinLeaseRentIntervalMonthly MinLeaseRentInterval = "monthly"
	MinLeaseRentIntervalWeekly  MinLeaseRentInterval = "weekly"
)

// Error defines model for Error.
type Error struct {
	Code    int32  `json:"code"`
	Message string `json:"message"`
	Type    string `json:"type"`
}

// ErrorResponse defines model for ErrorResponse.
type ErrorResponse struct {
	Error Error `json:"error"`
}

// GetLeaseRes defines model for GetLeaseRes.
type GetLeaseRes struct {
	Lease Lease `json:"lease"`
}

// GetTenantRes defines model for GetTenantRes.
type GetTenantRes struct {
	Tenant Tenant `json:"tenant"`
}

// Lease defines model for Lease.
type Lease struct {
	// Currency will default to USD when empty
	Currency     string             `json:"currency"`
	Deposit      int                `json:"deposit"`
	EndDate      openapi_types.Date `json:"endDate"`
	Id           string             `json:"id"`
	PropertyID   string             `json:"propertyID"`
	RentAmount   *int               `json:"rentAmount,omitempty"`
	RentInterval LeaseRentInterval  `json:"rentInterval"`
	StartDate    openapi_types.Date `json:"startDate"`
	TenantIDs    *[]string          `json:"tenantIDs,omitempty"`
}

// LeaseRentInterval defines model for Lease.RentInterval.
type LeaseRentInterval string

// LeaseList defines model for LeaseList.
type LeaseList struct {
	Leases []Lease `json:"leases"`
}

// LeasePropertyReq defines model for LeasePropertyReq.
type LeasePropertyReq struct {
	Lease MinLease `json:"lease"`
}

// MinLease defines model for MinLease.
type MinLease struct {
	// Currency will default to USD when empty
	Currency     string               `json:"currency"`
	Deposit      int                  `json:"deposit"`
	EndDate      openapi_types.Date   `json:"endDate"`
	PropertyID   string               `json:"propertyID"`
	RentAmount   *int                 `json:"rentAmount,omitempty"`
	RentInterval MinLeaseRentInterval `json:"rentInterval"`
	StartDate    openapi_types.Date   `json:"startDate"`
	TenantIDs    *[]string            `json:"tenantIDs,omitempty"`
}

// MinLeaseRentInterval defines model for MinLease.RentInterval.
type MinLeaseRentInterval string

// MinTenant defines model for MinTenant.
type MinTenant struct {
	DlNum    string             `json:"dlNum"`
	DlState  string             `json:"dlState"`
	Dob      openapi_types.Date `json:"dob"`
	FullName string             `json:"fullName"`
	Phones   []Phone            `json:"phones"`
}

// NewProperty defines model for NewProperty.
type NewProperty struct {
	City   string `json:"city"`
	State  string `json:"state"`
	Street string `json:"street"`
	Zip    string `json:"zip"`
}

// Phone defines model for Phone.
type Phone struct {
	Desc   string `json:"desc"`
	Number string `json:"number"`
}

// Property defines model for Property.
type Property struct {
	City   string `json:"city"`
	Id     string `json:"id"`
	State  string `json:"state"`
	Street string `json:"street"`
	Zip    string `json:"zip"`
}

// PropertyList defines model for PropertyList.
type PropertyList struct {
	Properties []Property `json:"properties"`
}

// StoreTenantReq defines model for StoreTenantReq.
type StoreTenantReq struct {
	Tenant MinTenant `json:"tenant"`
}

// Tenant defines model for Tenant.
type Tenant struct {
	DlNum    string             `json:"dlNum"`
	DlState  string             `json:"dlState"`
	Dob      openapi_types.Date `json:"dob"`
	FullName string             `json:"fullName"`
	Id       string             `json:"id"`
	Phones   []Phone            `json:"phones"`
}

// TenantList defines model for TenantList.
type TenantList struct {
	Tenants []Tenant `json:"tenants"`
}

// LeasePropertyJSONRequestBody defines body for LeaseProperty for application/json ContentType.
type LeasePropertyJSONRequestBody = LeasePropertyReq

// AddPropertyJSONRequestBody defines body for AddProperty for application/json ContentType.
type AddPropertyJSONRequestBody = NewProperty

// StorePropertyJSONRequestBody defines body for StoreProperty for application/json ContentType.
type StorePropertyJSONRequestBody = NewProperty

// AddTenantJSONRequestBody defines body for AddTenant for application/json ContentType.
type AddTenantJSONRequestBody = StoreTenantReq

// StoreTenantJSONRequestBody defines body for StoreTenant for application/json ContentType.
type StoreTenantJSONRequestBody = StoreTenantReq